Relocate client "structs" folder
