+++ src/client/structs/new_user_struct.py
